refactor: renames and documentation re: responsive navbar update
